+++ service/cloud_accounts/service.go
@@ -3,6 +3,8 @@ package cloud_accounts
 import (
 	"context"
 	"fmt"
+
+	"github.com/RedisLabs/rediscloud-go-api/redis"
 )
 
 type Log interface {
@@ -21,26 +23,26 @@ type Task interface {
 	Wait(ctx context.Context, id string) error
 }
 
-type Api struct {
+type API struct {
 	client HttpClient
 	task   Task
 	logger Log
 }
 
-func NewApi(client HttpClient, task Task, logger Log) *Api {
-	return &Api{client: client, task: task, logger: logger}
+func NewAPI(client HttpClient, task Task, logger Log) *API {
+	return &API{client: client, task: task, logger: logger}
 }
 
 // Create will create a new Cloud Account and return the identifier of the new account.
-func (a *Api) Create(ctx context.Context, account CreateCloudAccount) (int, error) {
+func (a *API) Create(ctx context.Context, account CreateCloudAccount) (int, error) {
 	var response taskResponse
 	if err := a.client.Post(ctx, "cloud account", "/cloud-accounts", account, &response); err != nil {
 		return 0, err
 	}
 
-	a.logger.Printf("Waiting for task %s to finish creating the cloud account", response.TaskId)
+	a.logger.Printf("Waiting for task %s to finish creating the cloud account", response)
 
-	id, err := a.task.WaitForResourceId(ctx, response.TaskId)
+	id, err := a.task.WaitForResourceId(ctx, redis.StringValue(response.ID))
 	if err != nil {
 		return 0, err
 	}
@@ -49,7 +51,7 @@ func (a *Api) Create(ctx context.Context, account CreateCloudAccount) (int, erro
 }
 
 // Get will retrieve an existing Cloud Account.
-func (a *Api) Get(ctx context.Context, id int) (*CloudAccount, error) {
+func (a *API) Get(ctx context.Context, id int) (*CloudAccount, error) {
 	var response CloudAccount
 	if err := a.client.Get(ctx, fmt.Sprintf("retrieve cloud account %d", id), fmt.Sprintf("/cloud-accounts/%d", id), &response); err != nil {
 		return nil, err
@@ -59,7 +61,7 @@ func (a *Api) Get(ctx context.Context, id int) (*CloudAccount, error) {
 }
 
 // Update will update certain values of an existing Cloud Account.
-func (a *Api) Update(ctx context.Context, id int, account UpdateCloudAccount) error {
+func (a *API) Update(ctx context.Context, id int, account UpdateCloudAccount) error {
 	var response taskResponse
 	if err := a.client.Put(ctx, fmt.Sprintf("update cloud account %d", id), fmt.Sprintf("/cloud-accounts/%d", id), account, &response); err != nil {
 		return err
@@ -67,7 +69,7 @@ func (a *Api) Update(ctx context.Context, id int, account UpdateCloudAccount) er
 
 	a.logger.Printf("Waiting for cloud account %d to finish being updated", id)
 
-	err := a.task.Wait(ctx, response.TaskId)
+	err := a.task.Wait(ctx, redis.StringValue(response.ID))
 	if err != nil {
 		return fmt.Errorf("failed when updating account %d: %w", id, err)
 	}
@@ -76,7 +78,7 @@ func (a *Api) Update(ctx context.Context, id int, account UpdateCloudAccount) er
 }
 
 // Delete will delete an existing Cloud Account.
-func (a *Api) Delete(ctx context.Context, id int) error {
+func (a *API) Delete(ctx context.Context, id int) error {
 	var response taskResponse
 	if err := a.client.Delete(ctx, fmt.Sprintf("delete cloud account %d", id), fmt.Sprintf("/cloud-accounts/%d", id), &response); err != nil {
 		return err
@@ -84,7 +86,7 @@ func (a *Api) Delete(ctx context.Context, id int) error {
 
 	a.logger.Printf("Waiting for cloud account %d to finish being deleted", id)
 
-	if err := a.task.Wait(ctx, response.TaskId); err != nil {
+	if err := a.task.Wait(ctx, redis.StringValue(response.ID)); err != nil {
 		return fmt.Errorf("failed when deleting account %d: %w", id, err)
 	}