+++ service/subscriptions/service.go
@@ -3,6 +3,8 @@ package subscriptions
 import (
 	"context"
 	"fmt"
+
+	"github.com/RedisLabs/rediscloud-go-api/redis"
 )
 
 type Log interface {
@@ -21,26 +23,26 @@ type Task interface {
 	Wait(ctx context.Context, id string) error
 }
 
-type Api struct {
+type API struct {
 	client HttpClient
 	task   Task
 	logger Log
 }
 
-func NewApi(client HttpClient, task Task, logger Log) *Api {
-	return &Api{client: client, task: task, logger: logger}
+func NewAPI(client HttpClient, task Task, logger Log) *API {
+	return &API{client: client, task: task, logger: logger}
 }
 
-func (a *Api) Create(ctx context.Context, subscription CreateSubscription) (int, error) {
+func (a *API) Create(ctx context.Context, subscription CreateSubscription) (int, error) {
 	var task taskResponse
 	err := a.client.Post(ctx, "create subscription", "/subscriptions", subscription, &task)
 	if err != nil {
 		return 0, err
 	}
 
-	a.logger.Printf("Waiting for task %s to finish creating the subscription", task.TaskId)
+	a.logger.Printf("Waiting for task %s to finish creating the subscription", task)
 
-	id, err := a.task.WaitForResourceId(ctx, task.TaskId)
+	id, err := a.task.WaitForResourceId(ctx, redis.StringValue(task.ID))
 	if err != nil {
 		return 0, err
 	}
@@ -48,7 +50,7 @@ func (a *Api) Create(ctx context.Context, subscription CreateSubscription) (int,
 	return id, nil
 }
 
-func (a *Api) Delete(ctx context.Context, id int) error {
+func (a *API) Delete(ctx context.Context, id int) error {
 	var task taskResponse
 	err := a.client.Delete(ctx, fmt.Sprintf("delete subscription %d", id), fmt.Sprintf("/subscriptions/%d", id), &task)
 	if err != nil {
@@ -57,7 +59,7 @@ func (a *Api) Delete(ctx context.Context, id int) error {
 
 	a.logger.Printf("Waiting for subscription %d to finish being deleted", id)
 
-	err = a.task.Wait(ctx, task.TaskId)
+	err = a.task.Wait(ctx, redis.StringValue(task.ID))
 	if err != nil {
 		return err
 	}